chore: use i32 instead of i64

+++ src/wasm/point_int_2d.rs
@@ -8,14 +8,14 @@ use crate::wasm::vector_int_2d::VectorInt2D;
 /// 2D point defined in js
 #[wasm_bindgen]
 #[derive(Copy, Clone, Debug)]
-pub struct PointInt2D(Point2<i64>);
+pub struct PointInt2D(Point2<i32>);
 
 #[wasm_bindgen]
 impl PointInt2D {
     // Statics
     /// Creates a new point from given coordinates
     #[wasm_bindgen(constructor)]
-    pub fn new(x: i64, y: i64) -> PointInt2D {
+    pub fn new(x: i32, y: i32) -> PointInt2D {
         PointInt2D(Point2::new(x, y))
     }
 
@@ -46,53 +46,53 @@ impl PointInt2D {
 
     // Properties
     #[wasm_bindgen(getter)]
-    pub fn x(&self) -> i64 {
+    pub fn x(&self) -> i32 {
         self.0.x
     }
 
     #[wasm_bindgen(setter)]
-    pub fn set_x(&mut self, x: i64) {
+    pub fn set_x(&mut self, x: i32) {
         self.0.x = x;
     }
 
     #[wasm_bindgen(getter)]
-    pub fn y(&self) -> i64 {
+    pub fn y(&self) -> i32 {
         self.0.y
     }
 
     #[wasm_bindgen(setter)]
-    pub fn set_y(&mut self, y: i64) {
+    pub fn set_y(&mut self, y: i32) {
         self.0.y = y;
     }
 }
 
 // Conversions
-impl AsRef<Point2<i64>> for PointInt2D {
-    fn as_ref(&self) -> &Point2<i64> {
+impl AsRef<Point2<i32>> for PointInt2D {
+    fn as_ref(&self) -> &Point2<i32> {
         &self.0
     }
 }
 
-impl AsMut<Point2<i64>> for PointInt2D {
-    fn as_mut(&mut self) -> &mut Point2<i64> {
+impl AsMut<Point2<i32>> for PointInt2D {
+    fn as_mut(&mut self) -> &mut Point2<i32> {
         &mut self.0
     }
 }
 
-impl Borrow<Point2<i64>> for PointInt2D {
-    fn borrow(&self) -> &Point2<i64> {
+impl Borrow<Point2<i32>> for PointInt2D {
+    fn borrow(&self) -> &Point2<i32> {
         &self.0
     }
 }
 
-impl BorrowMut<Point2<i64>> for PointInt2D {
-    fn borrow_mut(&mut self) -> &mut Point2<i64> {
+impl BorrowMut<Point2<i32>> for PointInt2D {
+    fn borrow_mut(&mut self) -> &mut Point2<i32> {
         &mut self.0
     }
 }
 
-impl From<Point2<i64>> for PointInt2D {
-    fn from(value: Point2<i64>) -> Self {
+impl From<Point2<i32>> for PointInt2D {
+    fn from(value: Point2<i32>) -> Self {
         PointInt2D(value)
     }
 }
@@ -105,9 +105,9 @@ impl PartialEq for PointInt2D {
     }
 }
 
-impl PartialEq<Point2<i64>> for PointInt2D {
+impl PartialEq<Point2<i32>> for PointInt2D {
     #[inline]
-    fn eq(&self, other: &Point2<i64>) -> bool {
+    fn eq(&self, other: &Point2<i32>) -> bool {
         &self.0 == other
     }
 }
